Moved the Turn Tracker Tests from the turn tracker tester folder to the backEnd test folder. It still works.
