--- src/com/cs428/pandemic/backEnd/PlaceHolderClass_DeleteThisWhenYouPushCode.java
@@ -1,5 +1,0 @@
-package com.cs428.pandemic.backEnd;
-
-public class PlaceHolderClass_DeleteThisWhenYouPushCode {
-
-}
